mises à jour

+++ src/modele/Poussin.java
@@ -27,7 +27,7 @@ public class Poussin extends Volatile {
     System.out.println("je suis un poussin");
   }
 
-  public void seDeplacer(long tempsEnSeconde) {
+  public void seDeplacer(double tempsEnSeconde) {
     //aller à droite. Lorsqu'il atteint la limite de l'univers, il va à gauche (le booléen change le sens)
     // Le poussin a un déplacement limité y=0
     if (deplacementDroite) {
